Sync missing files from upstream
There was a mistake when we updated our fork to the upstream at the
commit e70eac45b8ae8cf5b4e8c75496005c7198387ee3. Many files were missing
and in this patch we copy all the files from the upstream at that
commit to this fork.

See the commit 184fb08e301685ab62949ed593018740d289836b of *this*
repository for the reference regarding the previous sync with the
upstream.